Fixed all warnings.
git-svn-id: https://svn.uni-konstanz.de/disy/Treetank/trunk/treetank@4452 e3ddb328-5bfe-0310-b762-aafcbcbd2528

+++ src/main/java/org/treetank/openbsd/CryptoNativeImpl.java
@@ -117,6 +117,6 @@ public final class CryptoNativeImpl implements ICrypto {
       final byte tank,
       final byte operation,
       final short length,
-      final long address);
+      final int address);
 
 }
